+++ src/proxy.rs
@@ -101,7 +101,10 @@ pub async fn connect_udp(
                 Ok(UdpSocket::Proxy(socket))
             }
             ProxyProtocol::Http => {
-                unimplemented!()
+                return Err(io::Error::new(
+                    io::ErrorKind::Unsupported,
+                    "HTTP proxy does not support UDP associate",
+                ));
             }
         },
         None => TokioUdpSocket::bind(local_addr).await.map(UdpSocket::Tokio),